requirements.txt
pandas>=2.0
streamlit>=1.30
pyyaml>=6.0
msoffcrypto-tool>=5.0.0
xlrd>=2.0.1
pymupdf>=1.23
easyocr>=1.7
matplotlib>=3.7

openpyxl>=3.1
Plotly>=5.22